Add organizer to calendar (.ics) file
There is a strange behaviour in Outlook 2013 when there is no organizer.
A button appears to update the event making the person updating it
automatically a organizer for all people to which the update is sent.
To prevent that we can add the organizer to the file explicitely and only
a button to save the date appears.